backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
weasyprint
requests
streamlit
pdfminer.six
rapidfuzz
docx